Changed decl syntax.
decl has a new, more clear syntax, capable to express mutability and
represent every kind of object in a clear and repeatable way.

Declarations like
decl x func(a int8)

are now expressed in the more clear form

decl func x(a int8)

and same is for mut.

This is a purely syntactical change and has no consequences on the AST, so ctrans and anything that depends on
the parser should still work unchanged (assuming that their syntax is modified according to the new syntax)