v0.3.0. Big changes in code generation: now expressions are translated to C using an SSA-like form. This allows the generated code to be way more flexible - but at the expense of readability. While the new code is way more "bloated" when build on debug settings, it gets as fast as the equivalent C code when compiled with optimizations on any reasonably new compiler (GCC, Clang, probably MSVC) because SSA variables are coalesced.
This new expression generation system should't change behaviour at all - if it does, report a bug.

The ABI for functions returning structures is now well defined - they now instead return void, and the return value is allocated on the stack of the caller and passed through an hidden pointer called "__rvoOut". While less efficient than using a registry for small structures, it provides a warranty of constant, well defined behaviour.
This obviously breaks ABI and requires a full rebuilding with version 0.3.0.